[Fix] form component 분리
